Refactor collisions: replace PairWiseCoulombCollision class with BinaryCollision class (#2057)
* Refactor collisions: replace PairWiseCoulombCollision class with BinaryCollision class

* Make BinaryCollision a templated class with the functor type as parameter

* Update README.md
